#5- Design domain: user entitiy, role enum, converter
A new role system has been added, and since I have already tried creating a new role table to give multiple role to one member when writing in the role db column, I implemented it in a different way.

Role information is stored as a combined string and implemented as a converter.

The new role system is written with the 'RoleType' enum.