#53-Refactor test display name
Admin account controllers are like other controllers.
There is a handler method that provides a view and a handler method that provides data.
Since they are together, we decided that it was not appropriate to write "View" in the title of the all-encompassing test class, so we removed it from other controller tests, but it remained here. I even cleaned it out.