#26 - AdminLTE component and implementing Todo List by using javascript
Implement a simple functional To-Do list using AdminLTE design elements.
Because list information is not stored in the backend or in the browser,
When I refresh the screen, everything disappears.
It was made very simply, so there are many elements that can be upgraded.

The file that should be included to make todo list
- layout-main-table.js
- todolist.js
- layout-card-todolist.html
- layout-main-table.html
- layout-main-table.th.xml
- layout-scripts.html